backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "uvicorn>=0.35.0",
]
